flatten span link attributes
